# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Store API
    description: Service for a complete store with users, products, carts, and orders
    version: 0.0.1
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /cart:
        get:
            tags:
                - store
            summary: getCart store
            operationId: store#getCart
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Cart'
                        required:
                            - id
                            - userID
                            - items
                            - totalAmount
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/StoreGetCartNotFoundResponseBody'
            schemes:
                - http
    /cart/item:
        post:
            tags:
                - store
            summary: addToCart store
            operationId: store#addToCart
            parameters:
                - name: AddToCartRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CartItem'
                    required:
                        - productID
                        - quantity
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Cart'
                        required:
                            - id
                            - userID
                            - items
                            - totalAmount
            schemes:
                - http
    /cart/item/{productID}:
        delete:
            tags:
                - store
            summary: removeFromCart store
            operationId: store#removeFromCart
            parameters:
                - name: productID
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Cart'
                        required:
                            - id
                            - userID
                            - items
                            - totalAmount
            schemes:
                - http
    /login:
        post:
            tags:
                - store
            summary: loginUser store
            description: Login a user and return a JWT token
            operationId: store#loginUser
            parameters:
                - name: LoginUserRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/StoreLoginUserRequestBody'
                    required:
                        - username
                        - password
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/StoreLoginUserResponseBody'
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
            schemes:
                - http
...
    /orders/{id}:
        get:
            tags:
                - store
            summary: getOrder store
            operationId: store#getOrder
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Order'
                        required:
                            - id
                            - userID
                            - items
                            - totalAmount
                            - status
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/StoreGetOrderNotFoundResponseBody'
            schemes:
                - http
    /products:
        get:
            tags:
                - store
            summary: listProducts store
            operationId: store#listProducts
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/Product'
            schemes:
                - http
        post:
            tags:
                - store
            summary: createProduct store
            operationId: store#createProduct
            parameters:
                - name: CreateProductRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/NewProduct'
                    required:
                        - name
                        - price
                        - inventory
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/Product'
                        required:
                            - id
                            - name
                            - price
                            - inventory
            schemes:
                - http
    /products/{id}:
        get:
            tags:
                - store
            summary: getProduct store
            operationId: store#getProduct
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Product'
                        required:
                            - id
                            - name
                            - price
                            - inventory
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/StoreGetProductNotFoundResponseBody'
            schemes:
                - http
    /users:
        get:
            tags:
                - store
            summary: getUserAll store
            operationId: store#getUserAll
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/User'
            schemes:
                - http
        post:
            tags:
                - store
            summary: createUser store
            operationId: store#createUser
            parameters:
                - name: CreateUserRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/NewUser'
                    required:
                        - username
                        - email
                        - firstName
                        - lastName
                        - password
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/User'
                        required:
                            - id
                            - username
                            - email
            schemes:
                - http
    /users/{id}:
        get:
            tags:
                - store
            summary: getUser store
            operationId: store#getUser
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/User'
                        required:
                            - id
                            - username
                            - email
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/StoreGetUserNotFoundResponseBody'
            schemes:
                - http
...
    /users/update:
        post:
            tags:
                - store
            summary: updateUser store
            operationId: store#updateUser
            parameters:
                - name: UpdateUserRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/UserUpdatePayload'
                    required:
                        - email
                        - firstName
                        - lastName
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/User'
                        required:
                            - id
                            - username
                            - email
            schemes:
                - http
definitions:
    Cart:
        title: Cart
        type: object
        properties:
            id:
                type: string
                description: Unique cart ID
                example: Nesciunt qui cumque natus quas.
            items:
                type: array
                items:
                    $ref: '#/definitions/CartItem'
                description: Items in the cart
                example:
                    - productID: Sint quia commodi odio laboriosam.
                      quantity: 306858953100216253
                    - productID: Sint quia commodi odio laboriosam.
                      quantity: 306858953100216253
            totalAmount:
                type: number
                description: Total amount of the cart
                example: 0.6356852567933207
                format: double
            userID:
                type: string
                description: ID of the user who owns the cart
                example: Dolorem distinctio.
        example:
            id: Quae culpa odio.
            items:
                - productID: Sint quia commodi odio laboriosam.
                  quantity: 306858953100216253
                - productID: Sint quia commodi odio laboriosam.
                  quantity: 306858953100216253
                - productID: Sint quia commodi odio laboriosam.
                  quantity: 306858953100216253
                - productID: Sint quia commodi odio laboriosam.
                  quantity: 306858953100216253
            totalAmount: 0.506424476544269
            userID: Esse autem delectus autem odio veniam.
        required:
            - id
            - userID
            - items
            - totalAmount
    CartItem:
        title: CartItem
        type: object
        properties:
            productID:
                type: string
                description: ID of the product
                example: Sed aut et perspiciatis sit dolorem.
            quantity:
                type: integer
                description: Quantity of the product
                example: 3037912408950545240
                format: int64
        example:
            productID: Totam reprehenderit quidem et veritatis non non.
            quantity: 6908865122743982408
        required:
            - productID
            - quantity
...
    Product:
        title: Product
        type: object
        properties:
            description:
                type: string
                description: Product description
                example: Placeat ut ut.
            id:
                type: string
                description: Unique product ID
                example: Dolor praesentium.
            inventory:
                type: integer
                description: Available inventory
                example: 1019156399898440114
                format: int64
            name:
                type: string
                description: Product name
                example: Delectus non at suscipit et officiis omnis.
            price:
                type: number
                description: Product price
                example: 0.3177849568592131
                format: double
        example:
            description: Dolorem voluptas molestiae.
            id: Ipsam numquam cumque sed laboriosam laudantium.
            inventory: 5855704577479907427
            name: Recusandae temporibus.
            price: 0.44443020831588176
        required:
            - id
            - name
            - price
            - inventory
    StoreGetCartNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: false
        description: getCart_not-found_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...
    StoreLoginUserRequestBody:
        title: StoreLoginUserRequestBody
        type: object
        properties:
            password:
                type: string
                description: Password for login
                example: Odio ipsam et quas.
            username:
                type: string
                description: Username for login
                example: Voluptatem rerum temporibus dolor cupiditate id.
        example:
            password: Libero dolores eum in magni provident sapiente.
            username: Debitis est.
        required:
            - username
            - password
    StoreLoginUserResponseBody:
        title: StoreLoginUserResponseBody
        type: object
        properties:
            token:
                type: string
                description: JWT token for the authenticated user
                example: Et dolorem et.
        example:
            token: Porro dolores aperiam reiciendis provident quasi.
...